MOTU: more debugging and tweaks. TimestampedBuffer: ffado_timestamp_t changed to double for now - MOTU needs the extra precision.
git-svn-id: http://subversion.ffado.org/ffado/trunk@498 2be59082-3212-0410-8809-b0798e1608f0

+++ libffado/src/libutil/TimestampedBuffer.h
@@ -31,7 +31,9 @@
 #include "../debugmodule/debugmodule.h"
 #include "libutil/ringbuffer.h"
 
-typedef float ffado_timestamp_t;
+//typedef float ffado_timestamp_t;
+//#define TIMESTAMP_FORMAT_SPEC "%14.3f"
+typedef double ffado_timestamp_t;
 #define TIMESTAMP_FORMAT_SPEC "%14.3f"
 // typedef int64_t ffado_timestamp_t;
 // #define TIMESTAMP_FORMAT_SPEC "%012lld"
